+++ src/elements/tooltip/tooltip.snapshot.spec.ts
@@ -1,0 +1,33 @@
+import { expect } from '@open-wc/testing';
+import { html } from 'lit/static-html.js';
+
+import { fixture, testA11yTreeSnapshot } from '../core/testing/private.js';
+
+import type { SbbTooltipElement } from './tooltip.component.js';
+import './tooltip.component.js';
+
+describe(`sbb-tooltip`, () => {
+  let wrapper: HTMLElement;
+  let element: SbbTooltipElement;
+
+  beforeEach(async () => {
+    wrapper = await fixture(html`
+      <div>
+        <button id="trigger">Button</button>
+        <sbb-tooltip trigger="trigger">Tooltip</sbb-tooltip>
+      </div>
+    `);
+    element = wrapper.querySelector('sbb-tooltip')!;
+    element.open();
+  });
+
+  it('DOM', async () => {
+    await expect(wrapper).dom.to.be.equalSnapshot({ ignoreAttributes: ['style'] });
+  });
+
+  it('Shadow DOM', async () => {
+    await expect(element).shadowDom.to.be.equalSnapshot();
+  });
+
+  testA11yTreeSnapshot();
+});